requirements.txt
Django==1.8.2
django-appconf==1.0.1
django-avatar==3.0.0
django-bootstrap-form==3.2
django-bootstrap-pagination==1.6.1
django-bootstrap3==7.0.1
django-crispy-forms==1.6.0
django-debug-toolbar==1.4
django-filer==1.1.1
django-filter
django-guardian==1.4.4
django-mptt==0.8.3
django-polymorphic==0.9.1
django-private-media==0.1.3
Django-Select2==5.8.4
django-tables2==1.1.6
six
django-user-accounts==1.3.1
djangorestframework==3.3.3
easy-thumbnails==2.3
gunicorn==19.4.5
jsonfield==1.0.3
Markdown==2.6.6
Pillow==3.1.1
pinax-eventlog==1.1.1
pinax-theme-bootstrap==7.5.0
pinax-webanalytics==2.0.1
pytz
sqlparse==0.1.19
Unidecode==0.4.19
Fabric==1.13.1
django-redis
mongoengine
requests
MySQL-python
